+++ src/structs.rs
@@ -748,7 +748,7 @@ where
 /// Boxed parser doesn't expose internal representation in it's type and allows to return
 /// different parsers in different conditional branches
 ///
-/// You can create it with a single argument `construct` macro:
+/// You can create it with a single argument `construct` macro or by using `boxed` annotation
 #[doc = include_str!("docs/boxed.md")]
 pub struct ParseBox<T> {
     /// Boxed inner parser