+++ Changelog.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## bpaf [0.7.10], bpaf_derive [0.3.5] - unreleased
+- improve error messages for typos like `-llvm` instead of `--llvm`
+- improve error messages when a flag is accepted by a command but not directly
+- allow to derive position bool
+- derive anywhere and boxed
+
 ## bpaf [0.7.9], bpaf_derive [0.3.4] - 2023-02-14
 - `ParseFailure::exit_code`
 - A way to specify custom usage in derive macro